Much work on templating, adding pagination to Teachers list, fixing header problems and options in studydetails.
git-svn-id: http://joomlacode.org/svn/biblestudy/BibleStudy/Trunk@393 a73faa7a-dd35-0410-af13-bd1c84eb1a9d

+++ administrator/admin/models/templateedit.php
@@ -63,12 +63,18 @@ class biblestudyModeltemplateedit extends JModel {
 	function delete() {
 		$cids = JRequest::getVar( 'cid', array(0), 'post', 'array' );
 		$row =& $this->getTable();
-
+		
 		if (count( $cids ))
 		{
 			foreach($cids as $cid) {
+				if ($cid == 1)
+					{$this->setError('You cannot delete the default template');
+					return false;
+					}
 				if (!$row->delete( $cid )) {
-					$this->setError( $row->getErrorMsg() );
+					if($cid == 1)
+					{$this->setError('You cannot delete the default template');}
+					else {$this->setError( $row->getErrorMsg() );}
 					return false;
 				}
 			}
